refactor(editor): extract selected rect widget (#12290)
