+++ blocksuite/affine/all/src/widgets/edgeless-selected-rect/index.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-widget-edgeless-selected-rect';
