+++ blocksuite/affine/blocks/root/src/edgeless/edgeless-root-block.ts
@@ -45,7 +45,6 @@ import { css, html } from 'lit';
 import { query } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import type { EdgelessSelectedRectWidget } from './components/rects/edgeless-selected-rect.js';
 import { EdgelessPageKeyboardManager } from './edgeless-keyboard.js';
 import type { EdgelessRootService } from './edgeless-root-service.js';
 import { isCanvasElement } from './utils/query.js';
@@ -131,13 +130,6 @@ export class EdgelessRootBlockComponent extends BlockComponent<
 
   get gfx() {
     return this.std.get(GfxControllerIdentifier);
-  }
-
-  get selectedRectWidget() {
-    return this.host.view.getWidget(
-      'edgeless-selected-rect',
-      this.host.id
-    ) as EdgelessSelectedRectWidget;
   }
 
   get slots() {
