+++ blocksuite/affine/blocks/root/src/edgeless/edgeless-root-spec.ts
@@ -4,7 +4,6 @@ import { literal, unsafeStatic } from 'lit/static-html.js';
 
 import { NOTE_SLICER_WIDGET } from './components/note-slicer/index.js';
 import { EDGELESS_DRAGGING_AREA_WIDGET } from './components/rects/edgeless-dragging-area-rect.js';
-import { EDGELESS_SELECTED_RECT_WIDGET } from './components/rects/edgeless-selected-rect.js';
 
 export const edgelessDraggingAreaWidget = WidgetViewExtension(
   'affine:page',
@@ -15,11 +14,6 @@ export const noteSlicerWidget = WidgetViewExtension(
   'affine:page',
   NOTE_SLICER_WIDGET,
   literal`${unsafeStatic(NOTE_SLICER_WIDGET)}`
-);
-export const edgelessSelectedRectWidget = WidgetViewExtension(
-  'affine:page',
-  EDGELESS_SELECTED_RECT_WIDGET,
-  literal`${unsafeStatic(EDGELESS_SELECTED_RECT_WIDGET)}`
 );
 
 export class EdgelessLocker extends LifeCycleWatcher {
