+++ blocksuite/affine/blocks/root/src/effects.ts
@@ -1,5 +1,3 @@
-import { EdgelessAutoCompletePanel } from './edgeless/components/auto-complete/auto-complete-panel.js';
-import { EdgelessAutoComplete } from './edgeless/components/auto-complete/edgeless-auto-complete.js';
 import {
   NOTE_SLICER_WIDGET,
   NoteSlicer,
@@ -8,10 +6,6 @@ import {
   EDGELESS_DRAGGING_AREA_WIDGET,
   EdgelessDraggingAreaRectWidget,
 } from './edgeless/components/rects/edgeless-dragging-area-rect.js';
-import {
-  EDGELESS_SELECTED_RECT_WIDGET,
-  EdgelessSelectedRectWidget,
-} from './edgeless/components/rects/edgeless-selected-rect.js';
 import {
   EdgelessRootBlockComponent,
   EdgelessRootPreviewBlockComponent,
@@ -36,13 +30,6 @@ function registerRootComponents() {
 }
 
 function registerMiscComponents() {
-  // Auto-complete components
-  customElements.define(
-    'edgeless-auto-complete-panel',
-    EdgelessAutoCompletePanel
-  );
-  customElements.define('edgeless-auto-complete', EdgelessAutoComplete);
-
   // Note and template components
   customElements.define(NOTE_SLICER_WIDGET, NoteSlicer);
 
@@ -51,21 +38,14 @@ function registerMiscComponents() {
     EDGELESS_DRAGGING_AREA_WIDGET,
     EdgelessDraggingAreaRectWidget
   );
-  customElements.define(
-    EDGELESS_SELECTED_RECT_WIDGET,
-    EdgelessSelectedRectWidget
-  );
 }
 
 declare global {
   interface HTMLElementTagNameMap {
     'affine-edgeless-root': EdgelessRootBlockComponent;
     'affine-edgeless-root-preview': EdgelessRootPreviewBlockComponent;
-    'edgeless-auto-complete-panel': EdgelessAutoCompletePanel;
-    'edgeless-auto-complete': EdgelessAutoComplete;
     'note-slicer': NoteSlicer;
     'edgeless-dragging-area-rect': EdgelessDraggingAreaRectWidget;
-    'edgeless-selected-rect': EdgelessSelectedRectWidget;
     'affine-page-root': PageRootBlockComponent;
   }
 }
