+++ blocksuite/affine/widgets/edgeless-selected-rect/src/auto-complete-panel.ts
