+++ blocksuite/affine/widgets/edgeless-selected-rect/src/edgeless-auto-complete.ts
