+++ blocksuite/affine/widgets/edgeless-selected-rect/src/index.ts
@@ -1,0 +1,1 @@
+export * from './edgeless-selected-rect';
