+++ blocksuite/affine/widgets/edgeless-selected-rect/src/resize-handles.ts
