+++ tsconfig.json
@@ -100,6 +100,7 @@
     { "path": "./blocksuite/affine/shared" },
     { "path": "./blocksuite/affine/widgets/drag-handle" },
     { "path": "./blocksuite/affine/widgets/edgeless-auto-connect" },
+    { "path": "./blocksuite/affine/widgets/edgeless-selected-rect" },
     { "path": "./blocksuite/affine/widgets/edgeless-toolbar" },
     { "path": "./blocksuite/affine/widgets/edgeless-zoom-toolbar" },
     { "path": "./blocksuite/affine/widgets/frame-title" },
